[PBW-3961][Chrome][FW] - PreMidPost - Looping after it replays
In this.onReplay() within ad_manager_controller.js we call
_triggerAdManagerCallback(this.EVENTS.REPLAY_REQUESTED); which in turn
invokes the event listener for this event within the specific ad
manager (in this case freewheel ad manager). The event listener for
EVENTS.REPLAY_REQUESTED was missing in freewheel.js. Adding the event
listener onRePlayRequested fixed the issue. Appears to also solve
PBI-1259